move codes for log encoding
